# Controllers/CreateScriptApi.cs
/*
 * Script Generator
 *
 * To get the template, product, and raw data from the database
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using ScriptGenerator.Attributes;
using ScriptGenerator.Models;

namespace ScriptGenerator.Controllers
{
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class CreateScriptApiController : ControllerBase
    { 
        /// <summary>
        /// Add a new template
        /// </summary>
        /// <remarks>Add a new template to the database</remarks>
        /// <param name="body">Template data to be added</param>
        /// <response code="201">Template added successfully</response>
        /// <response code="400">Invalid input</response>
        [HttpPost]
        [Route("//api/v1/create-template")]
        [ValidateModelState]
        [SwaggerOperation("AddTemplate")]
        public virtual IActionResult AddTemplate([FromBody]V1CreatetemplateBody body)
        { 
            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            throw new NotImplementedException();
        }
    }
}