# wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Script Generator",
    "description" : "To get the template, product, and raw data from the database",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5000/",
    "description" : "API services for Script Generator."
  } ],
  "paths" : {
...
    "/api/v1/create-template" : {
      "post" : {
        "tags" : [ "create script" ],
        "summary" : "Add a new template",
        "description" : "Add a new template to the database",
        "operationId" : "AddTemplate",
        "requestBody" : {
          "description" : "Template data to be added",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/v1_createtemplate_body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Template added successfully"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "v1_createtemplate_body" : {
        "type" : "object",
        "properties" : {
          "templateId" : {
            "type" : "string"
          },
          "templateName" : {
            "type" : "string"
          },
          "templateContent" : {
            "type" : "string"
          }
        }
      }
    }
  }
}